# docs/swagger.yaml
basePath: /
definitions:
  handlers.AuthRequest:
    properties:
      password:
        type: string
      userName:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      url:
        type: string
    type: object
  handlers.MessageRequest:
    properties:
      body:
        type: string
    type: object
  handlers.RegisterResponse:
    properties:
      id:
        type: string
      userName:
        type: string
    type: object
  models.Message:
    properties:
      body:
        type: string
      id:
        type: string
    required:
    - body
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a go sandbox.
  title: Lets go chat
  version: "1.0"
paths:
  /message:
    post:
      consumes:
      - application/json
      description: Create new message
      parameters:
      - description: some text
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.MessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
      summary: Create new message
      tags:
      - messages
  /messages:
    get:
      consumes:
      - application/json
      description: get list of message
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Message'
            type: array
      summary: List of messages
      tags:
      - messages
  /user:
    post:
      consumes:
      - application/json
      description: Register
      parameters:
      - description: auth body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.RegisterResponse'
      summary: Register
      tags:
      - auth
  /user/active-users:
    get:
      consumes:
      - application/json
      description: get list of users
      produces:
      - application/json
      responses: {}
      summary: List of active users
      tags:
      - users
  /user/list:
    get:
      consumes:
      - application/json
      description: get list of users
      produces:
      - application/json
      responses: {}
      summary: List of users
      tags:
      - users
  /user/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: auth body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
      summary: Login
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"